Descriptor Cache, Reverse-Z, Texture and Model Managers

+++ Source/Util/Files.cpp
@@ -1,17 +1,17 @@
 /*
- *    Copyright 2023 - 2024 Rachit Khandelwal
+ * Copyright (c) 2023 - 2025 Rachit
  *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 #include "Files.h"
@@ -20,12 +20,10 @@
 #include <filesystem>
 #include <fstream>
 
-// Namespace aliases
 namespace filesystem = std::filesystem;
 
 namespace Engine::Files
 {
-    // Assets directory
     constexpr auto ASSETS_DIRECTORY = "Assets/";
 
     std::string GetAssetPath(const std::string_view prefix, const std::string_view fileName)
@@ -44,7 +42,7 @@ namespace Engine::Files
     {
         // This should always work
         static_assert(sizeof(usize) >= sizeof(std::uintmax_t), "How???");
-        return filesystem::file_size(path);
+        return static_cast<usize>(filesystem::file_size(path));
     }
 
     std::vector<u8> ReadBytes(const std::string_view path)
@@ -57,7 +55,6 @@ namespace Engine::Files
             Logger::Error("Failed to load shader binary {}!\n", path);
         }
 
-        // Pre-allocate memory for SPEED
         std::vector<u8> binary = {};
         binary.resize(GetFileSize(path));
